[dx] move sidebar settings from layout to sidebar
